+++ protected/extensions/eauth/EAuth.php
@@ -3,7 +3,7 @@
  * EAuth class file.
  *
  * @author Maxim Zemskov <nodge@yandex.ru>
- * @link http://code.google.com/p/yii-eauth/
+ * @link http://github.com/Nodge/yii-eauth/
  * @license http://www.opensource.org/licenses/bsd-license.php
  */
 
@@ -12,38 +12,45 @@
  * @package application.extensions.eauth
  */
 class EAuth extends CApplicationComponent {
-	
+
 	/**
 	 * @var array Authorization services and their settings.
 	 */
 	public $services = array();
-	
+
 	/**
 	 * @var boolean Whether to use popup window for the authorization dialog.
 	 */
 	public $popup = true;
-	
-public function init() {
-		parent::init();
 
-		$this->services['google_oauth']['class'] = 'CustomGoogleService';
-		
-		/*$this->services['twitter']['class'] = 'CustomTwitterService';
+	/**
+	 * @var mixed Cache component name to use. False to disable cache.
+	 */
+	public $cache = 'cache';
 	
-		$this->services['facebook']['class'] = 'CustomFBService';*/
+	/**
+	 * @var integer the number of seconds in which the cached value will expire. 0 means never expire.
+	 */
+	public $cacheExpire = 0;
 	
-		$this->services['vkontakte']['class'] = 'CustomVKService';
-	}
+	/**
+	 * @var string popup redirect view with custom js code
+	 */
+	protected $redirectView = 'redirect';
 
 	/**
 	 * Returns services settings declared in the authorization classes.
-	 * For perfomance reasons it uses Yii::app()->cache to store settings array.
+	 * For perfomance reasons it uses cache to store settings array.
 	 * @return array services settings.
 	 */
 	public function getServices() {
-		if (Yii::app()->hasComponent('cache'))
-			$services = Yii::app()->cache->get('EAuth.services');
-		if (!isset($services) || !is_array($services)) {
+		$services = false;
+		if (!empty($this->cache) && Yii::app()->hasComponent($this->cache)) {
+			$cache = Yii::app()->getComponent($this->cache);
+			$services = $cache->get('EAuth.services');
+		}
+
+		if (false === $services || !is_array($services)) {
 			$services = array();
 			foreach ($this->services as $service => $options) {
 				$class = $this->getIdentity($service);
@@ -54,12 +61,12 @@ public function init() {
 					'jsArguments' => $class->getJsArguments(),
 				);
 			}
-			if (Yii::app()->hasComponent('cache'))
-				Yii::app()->cache->set('EAuth.services', $services);
+			if (isset($cache))
+				$cache->set('EAuth.services', $services, $this->cacheExpire);
 		}
 		return $services;
 	}
-	
+
 	/**
 	 * Returns the settings of the service.
 	 * @param string $service the service name.
@@ -72,17 +79,17 @@ public function init() {
 			throw new EAuthException(Yii::t('eauth', 'Undefined service name: {service}.', array('{service}' => $service)), 500);
 		return $services[$service];
 	}
-	
+
 	/**
 	 * Returns the type of the service.
 	 * @param string $service the service name.
-	 * @return string the service type. 
+	 * @return string the service type.
 	 */
 	public function getServiceType($service) {
 		$service = $this->getService($service);
 		return $service->type;
 	}
-	
+
 	/**
 	 * Returns the service identity class.
 	 * @param string $service the service name.
@@ -93,7 +100,7 @@ public function init() {
 		if (!isset($this->services[$service]))
 			throw new EAuthException(Yii::t('eauth', 'Undefined service name: {service}.', array('{service}' => $service)), 500);
 		$service = $this->services[$service];
-		
+
 		$class = $service['class'];
 		$point = strrpos($class, '.');
 		// if it is yii path alias
@@ -106,24 +113,34 @@ public function init() {
 		$identity->init($this, $service);
 		return $identity;
 	}
-	
+
 	/**
-	 * Redirects to url. If the authorization dialog opened in the popup window, 
+	 * Change default redirect view to custom. Allow Yii alias.
+	 * @param string $view new name of view with js code
+	 */
+	public function setRedirectView($view)
+	{
+		$this->redirectView = $view;
+	}
+
+	/**
+	 * Redirects to url. If the authorization dialog opened in the popup window,
 	 * it will be closed instead of redirect. Set $jsRedirect=true if you want
 	 * to redirect anyway.
 	 * @param mixed $url url to redirect. Can be route or normal url. See {@link CHtml::normalizeUrl}.
 	 * @param boolean $jsRedirect whether to use redirect while popup window is used. Defaults to true.
 	 */
-	public function redirect($url, $jsRedirect = true) {		
-		require_once dirname(__FILE__).DIRECTORY_SEPARATOR.'EAuthRedirectWidget.php';				
+	public function redirect($url, $jsRedirect = true) {
+		require_once dirname(__FILE__).DIRECTORY_SEPARATOR.'EAuthRedirectWidget.php';
 		$widget = Yii::app()->getWidgetFactory()->createWidget($this, 'EAuthRedirectWidget', array(
 			'url' => CHtml::normalizeUrl($url),
 			'redirect' => $jsRedirect,
+			'view' => $this->redirectView
 		));
 		$widget->init();
 		$widget->run();
 	}
-	
+
 	/**
 	 * Simple wrapper for {@link CController::widget} function for render the {@link EAuthWidget} widget.
 	 * @param array $properties the widget properties.
@@ -135,7 +152,7 @@ public function init() {
 		$widget->init();
 		$widget->run();
 	}
-	
+
 	/**
 	 * Serialize the identity class.
 	 * @param EAuthServiceBase $identity the class instance.
@@ -144,7 +161,7 @@ public function init() {
 	public function toString($identity) {
 		return serialize($identity);
 	}
-	
+
 	/**
 	 * Serialize the identity class.
 	 * @param string $identity serialized value.
@@ -157,7 +174,7 @@ public function init() {
 
 /**
  * The EAuthException exception class.
- * 
+ *
  * @author Maxim Zemskov <nodge@yandex.ru>
  * @package application.extensions.auth
  * @version 1.0