+++ protected/extensions/eauth/EAuthUserIdentity.php
@@ -3,7 +3,7 @@
  * EAuthUserIdentity class file.
  *
  * @author Maxim Zemskov <nodge@yandex.ru>
- * @link http://code.google.com/p/yii-eauth/
+ * @link http://github.com/Nodge/yii-eauth/
  * @license http://www.opensource.org/licenses/bsd-license.php
  */
 
@@ -12,18 +12,19 @@
  * @package application.extensions.eauth
  */
 class EAuthUserIdentity extends CBaseUserIdentity {
+
 	const ERROR_NOT_AUTHENTICATED = 3;
 
 	/**
 	 * @var EAuthServiceBase the authorization service instance.
 	 */
 	protected $service;
-	
+
 	/**
 	 * @var string the unique identifier for the identity.
 	 */
 	protected $id;
-	
+
 	/**
 	 * @var string the display name for the identity.
 	 */
@@ -33,23 +34,22 @@ class EAuthUserIdentity extends CBaseUserIdentity {
 	protected $email;
 	protected $mobilePhone;
 	protected $homePhone;
-	
+
 	/**
 	 * Constructor.
 	 * @param EAuthServiceBase $service the authorization service instance.
 	 */
 	public function __construct($service) {
 		$this->service = $service;
-		$this->serviceName = $service->name;
 	}
-	
+
 	/**
 	 * Authenticates a user based on {@link service}.
 	 * This method is required by {@link IUserIdentity}.
 	 * @return boolean whether authentication succeeds.
 	 */
-	public function authenticate() {		
-	if ($this->service->isAuthenticated) {
+	public function authenticate() {
+		if ($this->service->isAuthenticated) {
 			$this->id = $this->service->id;
 			$this->name = $this->service->getAttribute('name');
 			
@@ -57,16 +57,17 @@ class EAuthUserIdentity extends CBaseUserIdentity {
 			$this->email = $this->service->getAttribute('email');
 			$this->mobilePhone = $this->service->getAttribute('mobilePhone');
 			$this->homePhone = $this->service->getAttribute('homePhone');
-			
-			/*$this->setState('id', $this->id);
+
+			$this->setState('id', $this->id);
+			$this->setState('name', $this->name);
 			$this->setState('service', $this->service->serviceName);
-			
-			$this->setState('firstName', $this->firstName);
-			$this->setState('email', $this->email);
-			$this->setState('mobile_phone', $this->mobilePhone);
-			$this->setState('home_phone', $this->homePhone);*/
-			
-			$this->errorCode = self::ERROR_NONE;		
+
+			// You can save all given attributes in session.
+			//$attributes = $this->service->getAttributes();
+			//$session = Yii::app()->session;
+			//$session['eauth_attributes'][$this->service->serviceName] = $attributes;
+
+			$this->errorCode = self::ERROR_NONE;
 		}
 		else {
 			$this->errorCode = self::ERROR_NOT_AUTHENTICATED;