+++ protected/extensions/eauth/EAuthWidget.php
@@ -3,7 +3,7 @@
  * EAuthWidget class file.
  *
  * @author Maxim Zemskov <nodge@yandex.ru>
- * @link http://code.google.com/p/yii-eauth/
+ * @link http://github.com/Nodge/yii-eauth/
  * @license http://www.opensource.org/licenses/bsd-license.php
  */
 
@@ -12,23 +12,28 @@
  * @package application.extensions.eauth
  */
 class EAuthWidget extends CWidget {
-		
+
 	/**
 	 * @var string EAuth component name.
 	 */
 	public $component = 'eauth';
-	
+
 	/**
 	 * @var array the services.
-	 * @see EAuth::getServices() 
+	 * @see EAuth::getServices()
 	 */
 	public $services = null;
-	
+
+	/**
+	 * @var array predefined services. If null then use all services. Default is null.
+	 */
+	public $predefinedServices = null;
+
 	/**
 	 * @var boolean whether to use popup window for authorization dialog. Javascript required.
 	 */
 	public $popup = null;
-	
+
 	/**
 	 * @var string the action to use for dialog destination. Default: the current route.
 	 */
@@ -42,28 +47,38 @@ class EAuthWidget extends CWidget {
 	 */
 	public function init() {
 		parent::init();
-		
+
 		// EAuth component
-		$component = Yii::app()->{$this->component};
-		
+		$component = Yii::app()->getComponent($this->component);
+
 		// Some default properties from component configuration
 		if (!isset($this->services))
 			$this->services = $component->getServices();
+
+		if (is_array($this->predefinedServices)) {
+			$_services = array();
+			foreach ($this->predefinedServices as $_serviceName) {
+				if (isset($this->services[$_serviceName]))
+					$_services[] = $this->services[$_serviceName];
+			}
+			$this->services = $_services;
+		}		
+		
 		if (!isset($this->popup))
 			$this->popup = $component->popup;
-		
+
 		// Set the current route, if it is not set.
 		if (!isset($this->action))
 			$this->action = Yii::app()->urlManager->parseUrl(Yii::app()->request);
 	}
-	
+
 	/**
 	 * Executes the widget.
 	 * This method is called by {@link CBaseController::endWidget}.
 	 */
     public function run() {
 		parent::run();
-		
+
 		$this->registerAssets();
 		$this->render('auth', array(
 			'id' => $this->getId(),
@@ -71,7 +86,7 @@ class EAuthWidget extends CWidget {
 			'action' => $this->action,
 		));
     }
-	
+
 	/**
 	 * Register CSS and JS files.
 	 */