+++ protected/extensions/eauth/custom_services/CustomGoogleService.php
@@ -1,50 +1,26 @@
 <?php
-/**********************************************************************************************
-*                            CMS Open Real Estate
-*                              -----------------
-*	version				:	1.3.1
-*	copyright			:	(c) 2012 Monoray
-*	website				:	http://www.monoray.ru/
-*	contact us			:	http://www.monoray.ru/contact
-*
-* This file is part of CMS Open Real Estate
-*
-* Open Real Estate is free software. This work is licensed under a GNU GPL.
-* http://www.gnu.org/licenses/old-licenses/gpl-2.0.html
-*
-* Open Real Estate is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
-* Without even the implied warranty of  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
-***********************************************************************************************/
+/**
+ * An example of extending the provider class.
+ *
+ * @author Maxim Zemskov <nodge@yandex.ru>
+ * @link http://github.com/Nodge/yii-eauth/
+ * @license http://www.opensource.org/licenses/bsd-license.php
+ */
 
-require_once dirname(dirname(__FILE__)).'/services/GoogleOAuthService.php';
+require_once dirname(dirname(__FILE__)).'/services/GoogleOpenIDService.php';
 
-class CustomGoogleService extends GoogleOAuthService {	
-	protected $jsArguments = array('popup' => array('width' => 750, 'height' => 450));
-	protected $scope = 'https://www.googleapis.com/auth/userinfo.profile+https://www.googleapis.com/auth/userinfo.email';
-	protected $client_id = '';
-	protected $client_secret = '';
-	protected $providerOptions = array(
-		'authorize' => 'https://accounts.google.com/o/oauth2/auth',
-		'access_token' => 'https://accounts.google.com/o/oauth2/token',
+class CustomGoogleService extends GoogleOpenIDService {
+	
+	//protected $jsArguments = array('popup' => array('width' => 450, 'height' => 450));
+	
+	protected $requiredAttributes = array(
+		'name' => array('firstname', 'namePerson/first'),
+		'lastname' => array('lastname', 'namePerson/last'),
+		'email' => array('email', 'contact/email'),
+		'language' => array('language', 'pref/language'),
 	);
 	
-	public function __construct() {
-		$this->title = Yii::t('labels', 'Google');
-	}
-	
 	protected function fetchAttributes() {
-		$info = (array)$this->makeSignedRequest('https://www.googleapis.com/oauth2/v1/userinfo');
-				
-		$this->attributes['id'] = $info['id'];
-		$this->attributes['name'] = $info['name'];
-		
-		if (!empty($info['link']))
-			$this->attributes['url'] = $info['link'];
-				
-		$this->attributes['id'] = $info['id'];
-		$this->attributes['firstName'] = $info['given_name'];
-		$this->attributes['email'] = (isset($info['verified_email']) && $info['verified_email']) ? $info['email'] : '';
-		$this->attributes['mobilePhone'] = '';
-		$this->attributes['homePhone'] = '';
+		$this->attributes['fullname'] = $this->attributes['name'].' '.$this->attributes['lastname'];
 	}
 }