+++ protected/extensions/eauth/custom_services/CustomTwitterService.php
@@ -1,45 +1,26 @@
 <?php
-/**********************************************************************************************
-*                            CMS Open Real Estate
-*                              -----------------
-*	version				:	1.3.1
-*	copyright			:	(c) 2012 Monoray
-*	website				:	http://www.monoray.ru/
-*	contact us			:	http://www.monoray.ru/contact
-*
-* This file is part of CMS Open Real Estate
-*
-* Open Real Estate is free software. This work is licensed under a GNU GPL.
-* http://www.gnu.org/licenses/old-licenses/gpl-2.0.html
-*
-* Open Real Estate is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
-* Without even the implied warranty of  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
-***********************************************************************************************/
-
+/**
+ * An example of extending the provider class.
+ *
+ * @author Maxim Zemskov <nodge@yandex.ru>
+ * @link http://github.com/Nodge/yii-eauth/
+ * @license http://www.opensource.org/licenses/bsd-license.php
+ */
+ 
 require_once dirname(dirname(__FILE__)).'/services/TwitterOAuthService.php';
 
 class CustomTwitterService extends TwitterOAuthService {	
-	protected $jsArguments = array('popup' => array('width' => 750, 'height' => 450));
-	protected $key = '';
-	protected $secret = '';
-	protected $providerOptions = array(
-		'request' => 'https://api.twitter.com/oauth/request_token',
-		'authorize' => 'https://api.twitter.com/oauth/authorize',  //'https://api.twitter.com/oauth/authenticate', 
-		'access' => 'https://api.twitter.com/oauth/access_token',
-	);
-	
-	public function __construct() {
-		$this->title = tt('twitter_label', 'socialauth');
-	}
 	
 	protected function fetchAttributes() {
 		$info = $this->makeSignedRequest('https://api.twitter.com/1/account/verify_credentials.json');
 	
 		$this->attributes['id'] = $info->id;
-		$this->attributes['firstName'] = $info->name; // $info->screen_name;
-		$this->attributes['email'] = '';
-		$this->attributes['mobilePhone'] = '';
-		$this->attributes['homePhone'] = '';
+		$this->attributes['name'] = $info->name;
 		$this->attributes['url'] = 'http://twitter.com/account/redirect_by_id?id='.$info->id_str;
+
+		$this->attributes['username'] = $info->screen_name;
+		$this->attributes['language'] = $info->lang;
+		$this->attributes['timezone'] = timezone_name_from_abbr('', $info->utc_offset, date('I'));
+		$this->attributes['photo'] = $info->profile_image_url;
 	}
 }