+++ protected/extensions/eauth/services/MoikrugOAuthService.php
@@ -5,7 +5,7 @@
  * Register application: https://oauth.yandex.ru/client/my
  * 
  * @author Maxim Zemskov <nodge@yandex.ru>
- * @link http://code.google.com/p/yii-eauth/
+ * @link http://github.com/Nodge/yii-eauth/
  * @license http://www.opensource.org/licenses/bsd-license.php
  */
 
@@ -82,8 +82,8 @@ class MoikrugOAuthService extends EOAuth2Service {
 	 */
 	public function makeSignedRequest($url, $options = array(), $parseJson = true) {
 		if (!$this->getIsAuthenticated())
-			throw new CHttpException(401, 'Unable to complete the authentication because the required data was not received.');
-		
+			throw new CHttpException(401, Yii::t('eauth', 'Unable to complete the authentication because the required data was not received.', array('{provider}' => $this->getServiceTitle())));
+			
 		$options['query']['oauth_token'] = $this->access_token;
 		$result = $this->makeRequest($url, $options);
 		return $result;