+++ protected/views/layouts/main.php
@@ -77,25 +77,11 @@
  
 		<div class="modal-header">
 		    <a class="close" data-dismiss="modal">&times;</a>
-		    <h4>Modal header</h4>
+		    <h4><?php echo Yii::t('labels','Login form')?></h4>
 		</div>
 		 
 		<div class="modal-body">
 		    <p><?php $this->widget('application.modules.user.components.LoginWidget');?></p>
-		</div>
-		 
-		<div class="modal-footer">
-		    <?php $this->widget('bootstrap.widgets.TbButton', array(
-		        'type'=>'primary',
-		        'label'=>'Save changes',
-		        'url'=>'#',
-		        'htmlOptions'=>array('data-dismiss'=>'modal'),
-		    )); ?>
-		    <?php $this->widget('bootstrap.widgets.TbButton', array(
-		        'label'=>'Close',
-		        'url'=>'#',
-		        'htmlOptions'=>array('data-dismiss'=>'modal'),
-		    )); ?>
 		</div>
 		 
 		<?php $this->endWidget(); ?>